+++ curso-java-loiane/exercicios-aula17/src/com/leandro/cursojava/aula17/MediaDasNotas.java
@@ -9,7 +9,7 @@ public class MediaDasNotas {
         System.out.println("Digite a quantidade de notas que serao avaliadas: ");
         int quantidade = scanner.nextInt();
 
-        int soma = 0;
+        double soma = 0;
 
         int contador = 0;
